feat(contenttools): M11+M12 Phase 2 — derive_attributes + ProcessAIAsset wire-up
M11 (derive_attributes module):
- New module: derive_attributes with 4 normal modes (Faceted/Smooth/AreaWeighted/AngleWeighted) + 2 tangent modes (AreaWeighted/MikkTSpace)
- Vendored mikktspace for tangent reconstruction
- AssetPipeline::Config extended with derive_params + auto_derive_after_subdivide cascade
- ai_asset_pipeline_params POD extended with derive/subdivide fields (M11.7)
- CLI flags: --derive, --derive-normal-mode, --derive-tangent-mode, --derive-faceted-angle-deg, --no-auto-derive, --subdivide-scheme
- 8 unit tests pass

M12.1-M12.3 (decoder + scene serializer):
- New PackedMeshDecoder inverts Geometry.cpp encoder for all 4 elements_type variants (position_only/static_color/static_normal/static_normal_texture)
- SceneBlobReader extended with ReadSceneHeader (extracts scene_name + materials + lod_group_names)
- SceneBlobWriter extended with SerializeScene + GetSceneSize (byte-for-byte matches Geometry.cpp::pack_data)
- 9 unit tests pass: 4 round-trips + index widening + empty mesh + multi-mesh + truncated buffer + header extraction

M12.4-M12.6 (ProcessAIAsset wire-up):
- ProcessAIAsset now runs pipeline::Run when any opt-in flag is on (repair/remesh/subdivide/uvatlas/derive/collision)
- Default path (no flags) preserves byte-identical legacy output via ImportFbx → pack_data
- ConfigurePipelineFromPOD maps POD fields to pipeline::Config
- Matched allocator: CoTaskMemAlloc (Win) / malloc (other) for buffer free/realloc

Root-cause fix:
- Engine/Utilities/Math.h:42 unpack_to_unit_float return type was u32, silently truncating 0.5→0. Fixed to f32. Discovered via PackedMeshDecoder static_normal round-trip (32768 LE = correct bytes, but unpack_dir returned -1).

Verification:
- 13/13 ContentTools test binaries pass (105 total cases)
- Sponza.fbx end-to-end: 393 meshes / 3263 meshlets / 0 validation errors
- --derive path: 393 meshes / 2273 meshlets / 0 errors
- Default path byte-identical to legacy 2-arg path (verified on Sponza + gate-metal-bars)

Known limitation (pre-existing, out of scope):
- --subdivide crashes with pmp::TopologyException on non-manifold input.
  Subdivide.cpp:150 to_pmp() is outside try/catch. Pipeline wiring is correct.

+++ Engine/Utilities/Math.h
@@ -39,7 +39,7 @@ namespace primal::math
 	}
 
 	template<u32 bits>
-	[[nodiscard]] constexpr u32 unpack_to_unit_float(u32 i)
+	[[nodiscard]] constexpr f32 unpack_to_unit_float(u32 i)
 	{
 		static_assert(bits <= sizeof(u32) * 8);
 		assert(i < ((u32)1 << bits));
